+++ ourhome-community/src/docs.asciidoc/index.adoc
@@ -1,0 +1,36 @@
+:hardbreaks:
+ifndef::snippets[]
+:snippets: ./build/generated-snippets
+endif::[]
+
+:root: /src/docs/asciidoc
+
+= OurHome Community API
+
+== User
+
+=== 회원가입
+'''
+
+`*Request*`
+
+include::{snippets}/user/sign-up/http-request.adoc[]
+include::{snippets}/user/sign-up/request-fields.adoc[]
+
+`*Response*`
+
+include::{snippets}/user/sign-up/http-response.adoc[]
+include::{snippets}/user/sign-up/response-fields.adoc[]
+
+=== 로그인
+'''
+
+`*Request*`
+
+include::{snippets}/user/login/http-request.adoc[]
+include::{snippets}/user/login/request-fields.adoc[]
+
+`*Response*`
+
+include::{snippets}/user/login/http-response.adoc[]
+include::{snippets}/user/login/response-fields.adoc[]
