+++ ourhome-community/src/main/java/com/armand/ourhome/community/controller/RestDocsController.java
@@ -1,0 +1,12 @@
+package com.armand.ourhome.community.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class RestDocsController {
+    @GetMapping("/")
+    public String renderRestDocs(){
+        return "index";
+    }
+}
