--- ourhome-community/src/test/java/com/armand/ourhome/community/Test.java
@@ -1,4 +1,0 @@
-package com.armand.ourhome.community;
-
-public class Test {
-}
