+++ ourhome-community/src/test/java/com/armand/ourhome/community/follow/controller/FollowControllerTest.java
@@ -1,0 +1,7 @@
+package com.armand.ourhome.community.follow.controller;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class FollowControllerTest {
+
+}
