+++ ourhome-community/src/test/java/com/armand/ourhome/community/follow/service/FollowServiceTest.java
@@ -1,0 +1,7 @@
+package com.armand.ourhome.community.follow.service;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class FollowServiceTest {
+
+}
